Refactor ADODB_mssqlnative::_fetchField()
The method contained duplicated logic to return the fieldObject(s) data,
once at the beginning of the function (if the data has already been
cached), and one at the end after initializing the cache.

Inverting the logic to start by initializing the data if it has not
already been cached avoids this.

Adjust PHPDoc return type.