Define _init() method in ADODB_pdo class
This is just an empty stub, which avoids having to define the method in
every child class, and ensures we don't get polymorphic calls

Removes empty stubs in child classes.

TODO: _init() does not seem to be called anywhere anymore. This used to
be done done in the _UpdatePDO() method, but it was removed as part of
the refactoring.